pyproject.toml
[project]
name = "smtp-smuggling-lab"
version = "0.0.0"
description = "SMTP smuggling reproduction lab and detection"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
    "dpkt>=1.9.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.23",
]

[tool.setuptools]
py-modules = []

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
